backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (MongoDB)
motor==3.3.2
beanie==1.25.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GitHub Integration
PyGithub==1.59.1
requests==2.31.0
aiohttp==3.9.1

# ML & Data Processing
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4

# Email
fastapi-mail==1.4.1

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Health Checks
healthcheck==1.3.3
